pyproject.toml
[project]
name = "regolith-pyrolysis-simulator"
version = "0.1.0"
description = "Oxygen Shuttle regolith refinery simulator for lunar and Mars feedstocks"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "flask-socketio>=5.3",
    "pyyaml>=6.0",
    "plotly>=5.18",
    "numpy>=1.26",
    "scipy>=1.12",
]

[project.optional-dependencies]
melts = ["PetThermoTools"]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
